requirements.txt
streamlit
numpy
pandas
plotly
gspread
google-auth
st-gsheets-connection
scipy